Move some tests to appropriate place
`AbstractTestDistributedQueries` is run for connectors than support
read/write operations.

- move read-only tests to `AbstractTestIntegrationSmokeTest`
- move connector independent tests to `AbstractTestEngineOnlyQueries`
